requirements.txt
python-telegram-bot
requests
groq
python-dotenv
